% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Takes a graph and an initial node, then calculates the shortest path from the initial node to every other node in the graph.}
\usage{
dijkstra(data, initial_node)
}
\arguments{
\item{data}{A data.frame with three variables (v1, v2 and w) that contains the nodes of the graph (from v1 to v2) with the length of the node (w)}

\item{initial_node}{The node to start with.}
}
\value{
The shortest path from the initial node to every other node.
}
\description{
Takes a graph and an initial node, then calculates the shortest path from the initial node to every other node in the graph.
}
\examples{
dijkstra(data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), 1)
dijkstra(data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)), 3)
}
\seealso{
[Dijkstra algorithm on wiki](https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm)
}
