\name{wiki_graph}
\docType{data}
\alias{wiki_graph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ wiki_graph dataframe
}
\description{
The data.frame with three variables v1, v2 and w that contains the edges of the graph fromv1 to v2 with the weight of the edge w }
\usage{
wiki_graph
}
\format{A data.frame containing three numeric columns.}
\source{\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
