\name{wiki_graph}
\alias{wiki_graph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ wiki_graph dataframe
}
\description{
The data.frame with three variables (v1, v2 and w) that contains the edges of the graph (fromv1 to v2) with the weight of the edge (w) }
\usage{
wiki_graph
}
%- maybe also 'usage' for other objects documented here.
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}

